Add compression dictionary clearing support
Currently compression dictionaries are cleared when either the "cache" or "cookies" are set in `Clear-Site-Data`.

But it would be very useful if there is a way to only clear the compression dictionaries without clearing HTTP caches and Cookies.

So, I'd like to propose a new "dictionaries" datatype in Clear-Site-Data HTTP Response Header Field.

https://github.com/WICG/compression-dictionary-transport/issues/34